// lib/api/generated/models/StorePageCardsBlock.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { StorePageCardItem } from './StorePageCardItem';
export type StorePageCardsBlock = {
    id: string;
    order?: number;
    enabled?: boolean;
    type: string;
    title?: (string | null);
    items?: Array<StorePageCardItem>;
};