// lib/api/generated/models/StorePageContactInfoBlock.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StorePageContactInfoBlock = {
    id: string;
    order?: number;
    enabled?: boolean;
    type: string;
    title?: (string | null);
    email?: (string | null);
    phone?: (string | null);
    whatsapp?: (string | null);
    address?: (string | null);
    opening_hours?: (string | null);
};