// lib/api/generated/models/StorePageFaqBlock.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { StorePageFaqItem } from './StorePageFaqItem';
export type StorePageFaqBlock = {
    id: string;
    order?: number;
    enabled?: boolean;
    type: string;
    title?: (string | null);
    items?: Array<StorePageFaqItem>;
};