// lib/api/generated/models/StorePageHeroBlock.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StorePageHeroBlock = {
    id: string;
    order?: number;
    enabled?: boolean;
    type: string;
    eyebrow?: (string | null);
    title: string;
    subtitle?: (string | null);
    image_url?: (string | null);
    image_alt?: (string | null);
    cta_label?: (string | null);
    cta_url?: (string | null);
};