// lib/api/generated/models/StorePageImageBlock.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StorePageImageBlock = {
    id: string;
    order?: number;
    enabled?: boolean;
    type: string;
    url: string;
    alt: string;
    caption?: (string | null);
    alignment?: 'left' | 'center' | 'right' | 'full';
};