// lib/api/generated/models/StorePageRichTextBlock.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StorePageRichTextBlock = {
    id: string;
    order?: number;
    enabled?: boolean;
    type: string;
    title?: (string | null);
    markdown: string;
};